refactor: organize std.lib files
