+++ code/deno/common/std.lib/mod.ts
@@ -1,0 +1,13 @@
+export * from './u.Cmd.ts';
+export * from './u.Fs.ts';
+export * from './u.Path.ts';
+export * from './u.Testing.ts';
+
+export * as DateTime from 'https://deno.land/std@0.224.0/datetime/mod.ts';
+export * as dotenv from 'https://deno.land/std@0.224.0/dotenv/mod.ts';
+export { color as c, color };
+
+/**
+ * Console: formatting colors.
+ */
+import * as color from 'https://deno.land/std@0.224.0/fmt/colors.ts';
