+++ code/deno/common/std.lib/u.Fs.ts
@@ -1,6 +1,7 @@
 import { exists } from 'jsr:@std/fs@1.0.3';
-import { Path } from './lib.Fs.Path.ts';
-import { copyDir, glob } from './lib.Fs.u.ts';
+
+import { Path } from './u.Path.ts';
+import { copyDir, glob } from './u.Fs.u.ts';
 
 export { Path };
